Restore trace files that were previously in LFS
LFS is the proper approach to store binary archives. However GitHub limits
this to 1gb for free, for all repositories owned by a user. Any fork by any
other user consumes this bandwidth. This makes LFS unattractive for public
repositories as LFS stops working if they become popular, unless the user
pays per-GB prices. Its better then to store the files directly in git,
even if a poor practice.

Additionally, many users are unfamiliar with LFS and are surprised when
they need to install it to run the checked-in files. That then doesn't
work, because of the bandwidth restrictions.

There isn't a good central location to stash these all, e.g. to make
them a build-time dependency. So for now just being dumb and letting
them exist in git.


Former-commit-id: 9303c98de19d762cfafbbe8bf5b0279cc10330b7