Improve delay calculation for scheduled expiration
When the TimerWheel is used with the new `Scheduler` option, the delay to
the next expiration event has to be calculated. This was scanning wheels
and buckets sequentially (165 total) in the adjusted order and returning
the first time period found. This missed that a cascade operations may
occur, which could bring down an event that fires earlier. Thus the
bucket in the next wheel has to be peek'd at and the minimum delay
taken.

This doesn't change functionality, but merely improves the promptness
of the new active expiration feature (disabled by default).


Former-commit-id: a915558ec085cb77233a4b342e883e5cb85eacfa