Add request metadata to simulator (fixes #374)
The primitive long key is replaced with an AccessEvent type. This can
contain other metadata, like the entry's weight or miss penalty, which
could be used by an admission or eviction policy. The AdaptSize request
trace was added to demonstrate this for entry weights. Extending the
metadata for latency, enhancing policies, etc. is left for future work.

The policy's characteristics defines which features it supports and the
trace's characteristics define the constraints being applied. The
simulator will filter out any policy that does not meet the configured
trace's requirements.


Former-commit-id: 0263e986bee2ab91a77450b03062c9bf76a6f831